adding "SpheroidGrid" and switching to new raster structure
Rasters are now demoted to storage containers. All functions involving trasters will instead operate on tmany. Functions that previously relied on raster.grid for spatial awareness will instead require passing grid as a parameter. This has numerous advantages. It means we can subclass Grid and make use of method signatures to enforce requirements about the type of raster that's operated on. It means we can overload method signatures to express different approaches that leverage requirements about the type of raster. It also means grid.voronoi does not have to be stored as a shared_ptr, and can be initialized through constructor delegates.