applying similar organization to scratch/models/crust
File structure by category theory really proved out. 

It promises every file only considers either a single data structure or a homset between them, and every folder considers only a single namespace that corresponds to a single category. A category diagram is linked to in the README.md of the folder for a category. 

The only exception to the rule is where the homset maps to a primitive data type, or where there is a compelling reason to privatize attributes. In either case, the homset is represented as methods within the object itself. 

Dependencies can be easily reasoned with. If a change is made to an interface, the footprint of that change can be easily reasoned with. Functionality can be quickly navigated to. Namespaces boundaries are well defined mathematically. File boundaries are well defined mathematically. Beautiful category diagrams can be trivially constructed with online editors and saved in nonproprietary formats, and can be linked to in the README.md to give a visual overview of the folder structure and project architecture. Combined with localizing unit tests, it's easy to build up tests incrementally, one homset at a time. You can easily see at a glance from a file navigator which functionality has been covered by unit tests, and how far along the project is. Functionality that is not properly tested can simply be avoided when writing `#include`s and the compilation can still be trusted to work.