renaming: tmany->series
"series" further disambiguates from the namespace "many", and leaves open the option to expand the namespace to provide other similar data structures (such as for stack memory). "tmany" is advantageous since it's unique, and better implies similar functionality to "glm::tvec<>", and we do intend to borrow the mental framework of glm, but "glm::tvec<>" is not often used by glm developers. We originally intended for developers to use typedefs of tmany like "uints" or "uvec3s", but we more often find ourselves using tmany to avoid using typedefs in library code. "series" better implies we are trying to achieve functionality similar to "Series" in the python "pandas" module, which is true.