new category in orbit with abstraction over`properties`
The new category is in the parent namespace. A faithful functor maps it to `properties`, which is what we mean when we say it provides an abstraction over `properties`. The category is designed for mapping between data structures that represent orbits, including position-velocity tuples that allow for applying force to change the orbit (`State`) and classical orbital elements that allow for the stable simulation of static orbits across many revolutions (`Elements`)