first pass at refactor to optimize pressure/temperature relations
Pressure/temperature relations are currently implemented using lambdas. Some inferred properties use lambdas that call out to other lambdas. This kind of implementation results in massive amounts of indirection in memory, which places a penalty on performance. However, this implementation is intended to be run on large batches of data, so the solution is not ideal. A better solution to this issue was not previously sought since enough constraints were placed on the compound library that implementation was difficult. It was previously thought the parameter space of these functions could be traversed upon initialization to create compact approximations that existed in stack memory, but this solution was liable to cause issues with accuracy and the size of footprints in memory. A better solution is to store the pressure/temperature relation in a way that can be manipulated analytically, but this must be done in a way that is sufficiently generic. This commit attempts to address the issue by introducing a new way to represent relations. Each property is represented in a standard form that is sufficiently generic to represent existing expressions while also being compact within memory, and closed under that property's mixing law. So far only a GasPropertyStateRelation is created, which represents pressure/temperature relations involving the heat capacity, thermal conductivity, and viscosity of gases.