refactoring unit tests and test utils for gases to use distance metrics
Distance metrics require fewer assumptions to be made during implementation compared to equality overloads. The distances they return can be compared against arbitrary thresholds that are tailored to the problem. This in turn allows them to be packaged with the rest of the class, rather than kept in a separate file as a unit test utility function.