renaming: Piecewise->Boxed
"Boxed" implies the function is multiplied by the "box" function where box(x)=1 if lo<x<hi and 0 otherwise. It better describes the purpose of the former "Piecewise" class, since a piecewise function is technically only a function defined in terms of nonoverlapping pieces, whereas we want to represent our functions using sums of piecewise functions that are 0 for all but a given region.