consolidating moduli into new "Handbook" concept to allow completion
Completion is a multi step process that creates temporary tables. A local scope is needed to create completed tables, and it's appropriate to create this scope inside a function, but that function should ideally derive something. There are many tables to derive, so it's not appropriate to express derivations using output reference parameters, but there is no concept to convey the combination of all derivations as a return type, so there is need to introduce a concept to convey them. This concept is called a "Handbook", since it is serves a similar role as chemical handbooks in the real world. For now, Handbooks are limited in scope, so they only contain those properties that correlate to each other and therefore require a multistep process to derive. So far, only a handbook for elasticities has been created