promoting compound library to inc: it is finally done
Suitable values are now guaranteed to exist for all properties, either as published values or from estimation. These values are vetted by unit tests so as to fall within chemically plausible ranges. Satisfactory data structures now exist to describe all aspects of the library. Data structures exist for all relations that are closed under their correlations, and a GenericRelation exists for those relations that cannot fit this criteria.  The library is now being used as the foundation for another library, "mixtures", which is currently under development. This fits the minimum criteria needed for a library to be promoted from spike/ to inc/. It is done.