adding vector calculus for dymaxion grids, still fails due to grid
arrows that stitch together squares on the dymaxion grid are slightly longer, therefore the difference between values at the endpoints is naturally greater. We can confirm this is why there are artifacts in rasters for vector calculus operations such as divergence and laplacian. Gradient does not seem to be effected.