rename: store→index
"store" was an earlier name that was conceived to generalize existing paradigms that were used in tectonics.* prior to the advent of procedural things like grids. Many components of store, such as known:: or series::, have nothing to do with storage in the general case. Their only commonality is that something is indexed, and it occurs in series. Some uncertainty in naming still remains, since "series::" is already taken but intersects with the functionality of "known::"