reworking dymaxion:: to work with unmodified standard noise algorithms
The previous implementation assumed an origin at the southernmost corner of squares. This implementation assumes an origin at the westernmost corner of squares. This difference means that offsetting a vertex by +[1,1] will always produce a single valid vertex. Every nonpolar point on the sphere is now covered exactly once by traversing a neighborhood of [0,1]×[0,1] from every vertex. This should permit the use of dymaxion grids with standard noise algorithms that use this neighborhood, such as value noise or perlin noise