fixes to dymaxion::Projection
A number of unexplained workarounds for Projection are now addressed, unit tests pass, and there is a fix to the directions that x and y correspond to. There is still one outstanding workaround, which is that x and y for grid_id must still be swapped, it is uncertain why this is needed.