fixing hypsography
We were previously passing a single uniform variable to the CDF for earth's hypsography to accurately reproduce the distribution of terrain on earth. However we later switched to FBM, and this particular FBM was a linear combination of multiple independent uniformly distributed variables, so its PDF is not uniform, causing the hypsography of the world to not reflect what was asked for. The PDF for a linear combination of uniforms is not well studied, but it is known that it would be a spline, so finding the PDF is a difficult task, and calculating from the PDF would be expensive. Tectonics.js previously worked around this limitation by sorting the heights that were generated and scaling their index in the sorted list to the [0,1] interval, however we wanted to avoid using this approach again since it was slow and did not scale well.  Rather than try to find or approximate the PDF, we instead modified the FBM to act as a linear combination of Gaussian distributed variables. Since Gaussians are closed under addition and scalar multiplication, the PDF for this modified FBM is guaranteed to be Gaussian. This allows us to correctly model any hypsography desired by piping results from FBM to the Gaussian CDF and then to the inverse hypsographic CDF.