reworking part counts for `MineralStore` to better accommodate js models
models that are used within Tectonics.js are already well tested, and we would like to reuse them them wherever possible now that we are focusing upon MVP, however crust generation is simplified if we use the representation offered by`rock::Mineral`, and we intend to migrate to that representation anyway, so it also suites us to use anything within spike/model/rock that has already been implemented and thoroughly tested. The biggest risk that was introduced by the representation under spike/model/rock is the memory footprint it would require. We are already pressing against the limits of what we can model with ~40k cells, we would like to keep this resolution, but any further increases to memory footprint risks making the model intolerably slow, so we likely need to scale back our ambitions. The strata model proposed within spike/model/rock/strata was also high risk, since it was untested and it was uncertain as to whether it could adequately represent rock strata at a resolution that was sufficiently interesting while being performant. As a result, I've reevaluated the design, starting from first principles, and documented findings in inc/model/rock/README.md. We still use rock strata, but only to a limited extent, to represent functional distinctions that were represented within Tectonics.js but would otherwise have been unrepresented within Tectonics.cpp. These will soon be represented using new "bedrock", "sediment", and "sedimentary" layers. Since only the sediment layer is subject to spatial transport, we'll want to minimize the footprint of that out-of-order traversal, so these rock layers will be stored as separate rock::Crust objects. As a result, rock::Mineral and rock::Stratum will be reused with minimal modification, but rock::Strata and associated classes like rock::StrataValues will be discarded. The previous model under inc/model/rock/mineral was also extremely limited with respect to how metamorphic rock would appear in sediment layers - all mass that was associated with a single layer and compound could only assume a single phase, so we could not represent certain situations like alluvial deposits of precious stones like diamond or jade. This results in the first line of work, appearing in this commit, which adds a crude metamorphic/nonmetamorphic distinction to rock::Mineral so as to at least signal the possibility for such situations. This enhancement does not increase the size of rock::Mineral, but it does more effectively use its footprint while maintaining cache alignment. In order to make this possible, we've had to remove bins for mechanical weathering, however we still plan to represent mechanical weathering by coopting the otherwise unused "unweathered" bins within the sediment layer. Upon lithifying, this conglomerate rock will become representationally indistinguishable from homogeneous sedimentary rock, but this is presumed less interesting to the user compared to what metamorphic particles in sediment could provide.