dropping pressure and temperature records from stratum models
Tracking metamorphosis as separate mineral pools is necessary since sediment and sedimentary rock needs some way to represent metamorphosed particles and clasts, however if that is the only way to track metamorphosis, then any formation that has even a slight fraction over the threshold for metamorphosis will be entirely converted after only a few iterations of the simulation. To resolve this issue, pressure and temperature records were introduced to track metamorphosis for layers such as bedrock, but analysis has revealed that tracking metamorphosis this way will cause any mass that's added after metamorphic conditions subside to be incorrectly metamorphosed. The resolution we've found is to store dedicated formations to represent metasedimentary and metaigneous rock, independantly. This resolves all the above issues, and it no longer requires tracking pressure/temperature records