rename: spike ⟶ sketch
"spike" is a borrowing from rugby and suggests a short, simple experiment, typically involving brogrammers, which most definitely does not describe the contents of the folder. "sketch" is a borrowing from art and (from there) mathematics, and suggests that something is a rough outline that is not fully formed, but is needed for eventual completion. This better represents the intended usage of the folder, which is meant to contain code that outlines what the final product is supposed to look like and tries to be as close to executable C++ as possible, but is not ready to be relied on, and indeed may not even have undergone an attempt to compile.