discretizing statements on D to avoid trouble with integral notation
Subdivision into N segments is more of a thought experiment that allows us to determine the calculation for more complex geometry. The thought experiment asserts that contributions are equally distributed between segments, and that there's some resource between segments that can be allocated to make this the case. Algebraically, N goes away if we let the thought experiment take its course, as well as any concern for what the resource is. Integration requires there to be something infinitesimal that we're using as weights in a infinite linear combination, but since this infinitesimal is the unspecified resource in question, we're best left using discretized sums.