new grouped:: namespace in prep for floodfill/voronoi plate fracture
Voronoi requires seed positions, so for plate fracture to use it would require constructing those seeds from the plate id raster that floodfill provides, which is effectively an aggregate of vertex positions for each plate. This requires a set of "grouped" operations. This is behavior equivalent to `group.hpp` that was implemented earlier but not ever used and not ever transitioned to the new class-category style of index operations, so we've converted it here.