introducing `field::SphericalHarmonic`
Development was driven by the need for a scale-invariant technique that could smooth out scalar fields using an in-order traversal through memory. Smoothing is necessary to construct sensible looking plate boundaries. We had previously accomplished smoothing by repeatedly subtracting laplacians, but calculation of the laplacian required out-of-order traversal so that performance placed strict limits on the frequency of plate fracture. Smoothing using spherical harmonics may also potentially free us from using binary morphology when creating plate boundaries, which is both out-of-order and scale-dependent. The hope is that we will eventually have a technique that allows plate fracture to occur whenever forces are sufficient and not simply during rare global restructuring events as it is currently done.