introducing `dymaxion::VertexDownsamplingIds` with downsampling demo
Downsampling is needed now that we have demonstrated that Spherical Harmonics will not smooth vertex ids with the performance and accuracy that's needed for plate fracture. Tectonics.js addressed this problem using downsampling, but the performance of its implementation suffered since its grid lacked the ability to cleanly subdivide vertices during downsampling, largely because it did not subdivide the sphere into Cartesian grids. We previously got around this using nearest neighbor sampling using a lookup table for nearest neighbors. However Tectonics.cpp has addressed this issue by making it so that `dymaxion::Grid` cleanly subdivides the sphere into cartesian grids. Downsampling is now accomplished using a procedural series, `VertexDownsamplingIds`, since that can be combined by operations under `grouped::` to create downsampling logic of arbitrary complexity. To illustrate, using the same logic that's used for other grouping behavior, you could calculate the minimum, the product, standard deviation, etc. for all vertices in the fine grid that correspond to the same vertex of the coarse grid. We leverage this functionality within the downsampling/ demo when calculating averages for each vertex.