updating documentation for spheres and stars
also generalizing, simplifying, and introducing assumptions for their shaders. On further reflection, further generalization of these shaders (such as to account for atmosphere/photosphere scattering or specular reflection) is probably not needed given their exclusive intended use for being viewed at a distance, so this may be the last commit that touches this feature. This will change if a need is discovered, or implementation is found not to be too difficult.