decoupling model matrix and view
Plate motion will be accomplished by representing each plate with same grid in a different orientation. Therefore vertex positions in the gl buffer will be reused but the model matrix will change. It would be nice if we could reduce the number of draw calls by representing each plate as an instance, but each plate absolutely needs its own per-vertex attributes, and instancing in gl does not allow attributes that are unique to a combination of vertex and instance. Model matrices could be represented as a per-vertex attribute but the vast majority of attributes in this case would be the same. Model matrices are relatively large so this approach would be very wasteful. We therefore need to represent each plate with a separate draw() call. This is fine since each draw() call passes its own model matrix, however this model matrix was previously tied up with the rest of a ViewState object. It would be undesireable to create or modify a ViewState object with each draw call since creation is slow and modification poses risk associated with state management. In any case the model matrix has good reason not to belong in an object that represents concerns of the view. We therefore decouple model matrix from ViewState and pass it as a separate parameter in the draw() call.