documenting planned modification to dymaxion design
The new design should allow grid_id to be found using just 1 atan to find the square, then 2 dots and 1 atan for each of the coordinate lookups. The existing method requires more complex logic, which is difficult to summarize, but does require a cross product and inverse among other things.