fixing issue in Grid that caused segfault in unit test
The segfault was introduced when optimizing Grid by dropping a call to standardize() in a function that shouldn't have needed it since the grid id that was being standardized was already clamped to a range that would prevent a segfault. It's not yet clear to me why this call standardize() would be needed to prevent a segfault.