creating a new generalized alternative to FormationGenerationByElevation
FormationGenerationByElevation required a passed function to accept both an area and elevation, which made it difficult to compose with existing functions like field::compose(). FormationGeneration only requires that a passed function accepts a position. The result is scaled by area. This makes Its usage much more flexible. FormationGenerationByElevation could only realistically be used for model initialization, whereas FormationGeneration can be repurposed to specify rock that was newly created by rifting (which is why we're working on it now). EarthlikeIgneousFormationGeneration has already been converted to use the new FormationGeneration. This represents the most common use case but some usages of FormationGenerationByElevation likely remain. We plan to convert these next so that we can drop FormationGenerationByElevation entirely