CrustMotion no longer stores Grid as an attribute
CrustMotion was originally designed under the premise that attributes of Grid could be found in a fast and completely procedural way. Following from this assumption, the memory footprint of Grid would be very small, so it could be stored as an attribute with little consequence. Performance testing revealed that while this was generally true, some properties like vertex_dual_area had to be cached. This meant the footprint of Grid had to be much bigger, so we must move away from designs that store it as an attribute. This is generally okay though, since many *Grid classes may exist that aren't procedural, and the designs we move towards will better handle these cases while still working suitably well with procedural use cases.