fixing displacement display while rifting
Plate objects were modified so that 0 thickness indicated nonexistance. This was done to fix checks for existance like CrustSummaryPredicates::exists, but it broke other functionality like density display, since composition of CrustSummaries assumed real valued densities. This in turn broke display of displacment. The solution so far is to tweak the logic that sets min and max color values. Other options are on the table in the future.