slab motion is now calculated each frame
This should in principle alleviate issues where continents slide past each other. This causes a massive slowdown, but it should be easy to alleviate this by running the slab motion calculation in parallel. However it will need some work to make sure each processor has its own set of rasters to work with.