slab motion is now calculated in parallel
The effect of dynamic slab motion is now apparent. The instantaneous slab detachment is likely causing plate motion to stop prematurely. Real life slabs would likely continue to sink into the mantle well before detachment occurs, and while attached they would still impart a force on the rest of the plate. We could act as if certain subducted cells are sinking into the mantle without modeling their position correctly, but this could complicate logic since the same cells on the same plate might be needed to represent other unsubducted crust. This is why we've always chosen something similar to the current approach, where crust is destroyed shortly after foundering. A better way to implement deceleration caused by continent-continent collision might be to calculate the initial slab motion of a plate as was done previously but offset it each frame by a force imparted by continent-continent collisions.