Universals now work in python
Also, returning to old folder and namespace structure since correctly testing behavior will require working with all the contents of old orbit::, and orbital functionality was taking over the body:: namespace. The future of body:: is uncertain. I'll have to wait for the dust to settle to see what's a good organization for it.